requirements.txt
aiogram==3.4.1
fastapi==0.104.1
uvicorn==0.24.0
mercadopago==2.2.0
python-dotenv==1.0.0
aiofiles==23.2.1
python-multipart==0.0.6 